+++ Globals.py
@@ -1,10 +1,11 @@
 # Constants
-_CONFIG_FILE = 'config.xml'
 _VERSION = 'alpha-0.4.0'
 _TITLE = f"ToNCodes {_VERSION} - by MichelleDB"
+_GITHUB = 'https://github.com/69MichelleDB/ToNCodes'
+
+# Main window
 _WIDTH = 620
 _HEIGHT = 300
-_GITHUB = 'https://github.com/69MichelleDB/ToNCodes'
 
 # Options window
 _WIDTH_OPT = 550
@@ -14,10 +15,15 @@ _HEIGHT_OPT = 180
 _WIDTH_ABOUT = 500
 _HEIGHT_ABOUT = 200
 
+# Folders
 _FOLDER_CODES = 'Codes'
 _FOLDER_LOGS = 'Logs'
 _FOLDER_TEMPLATES = 'Templates'
 _FOLDER_TEMP = 'Temp'
+
+# Files
+_FILE_CONFIG = 'config.xml'
+_FILE_CONTROL = 'control.xml'
 
 # Global variables
 configList = {}