+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [WIP] - 2025..
 
 ### Changes
-- Fixed title.
+- Fixed main window's title.
+- Optimizations reading log files. [#8](https://github.com/69MichelleDB/ToNCodes/issues/8)
 
 ## [alpha-0.4.0] - 2025.02.14
 
@@ -34,4 +35,3 @@
 - Automatically finds ToN codes in VRChat's logs and saves them in .xml files inside the Codes folder.
 - You can copy a selected Code in the list by double clicking.
 - You can delete a selected Code by pressing the "Delete" key on your keyboard.
-